cgroup: split systemd code to a new file
