+++ src/libcrun/cgroup-systemd.h
@@ -1,0 +1,32 @@
+/*
+ * crun - OCI runtime written in C
+ *
+ * Copyright (C) 2017, 2018, 2019 Giuseppe Scrivano <giuseppe@scrivano.org>
+ * crun is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation; either version 2.1 of the License, or
+ * (at your option) any later version.
+ *
+ * crun is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with crun.  If not, see <http://www.gnu.org/licenses/>.
+ */
+#ifndef CGROUP_SYSTEMD_H
+#define CGROUP_SYSTEMD_H
+
+#include "container.h"
+#include <unistd.h>
+
+int libcrun_cgroup_enter_systemd (struct libcrun_cgroup_args *args,
+                                  struct libcrun_cgroup_status *out,
+                                  libcrun_error_t *err);
+int libcrun_destroy_systemd_cgroup_scope (struct libcrun_cgroup_status *cgroup_status,
+                                          libcrun_error_t *err);
+
+int parse_sd_array (char *s, char **out, char **next, libcrun_error_t *err);
+
+#endif
